+++ Chrominsky.Utils/Enums/SearchOrder.cs
@@ -1,0 +1,17 @@
+namespace Chrominsky.Utils.Enums;
+
+/// <summary>
+/// Represents the order in which a search operation should be performed.
+/// </summary>
+public enum SearchOrder
+{
+    /// <summary>
+    /// Indicates that the search operation should be performed in ascending order.
+    /// </summary>
+    Ascending,
+
+    /// <summary>
+    /// Indicates that the search operation should be performed in descending order.
+    /// </summary>
+    Descending,
+}