+++ Chrominsky.Utils/Repositories/Base/IBaseDatabaseRepository.cs
@@ -1,4 +1,4 @@
-using Chrominsky.Utils.Models.Base;
+using Chrominsky.Utils.Models;
 using Chrominsky.Utils.Models.Base.Interfaces;
 
 namespace Chrominsky.Utils.Repositories.Base;
@@ -48,4 +48,17 @@ public interface IBaseDatabaseRepository<T>
     /// <param name="id">Unique identifier of the entity.</param>
     /// <returns>True if the entity was successfully deleted, otherwise false.</returns>
     Task<bool> DeleteAsync<T>(Guid id) where T : class, IBaseDatabaseEntity;
+
+    /// <summary>
+    /// Performs a search operation on the database based on the provided search parameters.
+    /// </summary>
+    /// <typeparam name="T">Type of entity to be searched.</typeparam>
+    /// <param name="request">The search parameters and criteria.</param>
+    /// <returns>An enumerable collection of entities of type T that match the search criteria.</returns>
+    /// <remarks>
+    /// This method is intended to be used for searching entities based on specific conditions.
+    /// The search parameters are defined in the <see cref="SearchParameterRequest"/> class.
+    /// The returned collection may be empty if no matching entities are found.
+    /// </remarks>
+    Task<List<T>> SearchAsync<T>(SearchParameterRequest request) where T : class, IBaseDatabaseEntity;
 }