+++ Chrominsky.Utils/CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 
 This format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [1.0.6] - 2024-06-10
+### Fixed
+- `BaseDatabaseRepository` - fixed UpdatedAt not updating on `UpdateAsync` method
+
+### Added
+- `SearchParameterRequest` - added request structure for `SearchAsync` method in `IBaseDatabaseRepository`
+- `SearchParameter` - added parameter structure
+- `SearchOperator` - added search operator enum
+- `SearchOrder` - added search order enum
+- `ErrorMessage` - added error message structure
+- Unit tests project `Chrominsky.Utils.Tests` + few unit tests
+
+### Changed
+- `IBaseDatabaseRepository` - added `SearchAsync` method for supporting search functionality
+- `BaseDatabaseRepository` - added implementation of `SearchAsync` method
+
 ## [1.0.5] - 2024-06-05
 ### Changed
 - `BaseDatabaseRepository` - changed UpdateAsync to update only changed values leaving old ones intact